# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agrotech API",
    "description": "Api para gerenciamento de plantações e safras",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "OpenAPI Documentation - Development"
    }
  ],
...
  "paths": {
...
    "/api/fields/{id}": {
      "get": {
        "tags": [
          "Fields"
        ],
        "summary": "Fetch one field by its id",
        "operationId": "getFieldById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Field fetched",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FieldResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid UUID format"
          },
          "404": {
            "description": "Field not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "tags": [
          "Fields"
        ],
        "summary": "Update field",
        "operationId": "updateField",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FieldUpdateRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Field updated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FieldResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data or validation error"
          },
          "404": {
            "description": "Field not found"
          },
          "409": {
            "description": "Field title already exists for this property"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "Fields"
        ],
        "summary": "Delete field",
        "operationId": "deleteField",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Field deleted"
          },
          "400": {
            "description": "Invalid UUID format"
          },
          "404": {
            "description": "Field not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/fields": {
      "get": {
        "tags": [
          "Fields"
        ],
        "summary": "Fetch all fields with pagination",
        "operationId": "getAllFields",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "pagedAssembler",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PagedResourcesAssemblerFieldResponseDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fields fetched",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelFieldResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "Fields"
        ],
        "summary": "Create new field",
        "operationId": "createField",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FieldCreateRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Field created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FieldResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Property not found"
          },
          "409": {
            "description": "Field title already exists for this property"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/sensors/field/{fieldId}": {
      "get": {
        "tags": [
          "Sensors"
        ],
        "summary": "Fetch all sensors from a field",
        "operationId": "getSensorsByField",
        "parameters": [
          {
            "name": "fieldId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "pagedAssembler",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PagedResourcesAssemblerSensorResponseDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sensors fetched",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelSensorResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination parameters"
          },
          "404": {
            "description": "Field not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/harvests/field/{fieldId}": {
      "get": {
        "tags": [
          "Harvests"
        ],
        "summary": "Fetch all harvests from a field",
        "operationId": "getHarvestsByField",
        "parameters": [
          {
            "name": "fieldId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "pagedAssembler",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PagedResourcesAssemblerHarvestResponseDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Harvests fetched",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelHarvestResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination parameters"
          },
          "404": {
            "description": "Field not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/fields/property/{propertyId}": {
      "get": {
        "tags": [
          "Fields"
        ],
        "summary": "Fetch all fields from a property",
        "operationId": "getFieldsByProperty",
        "parameters": [
          {
            "name": "propertyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "pagedAssembler",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PagedResourcesAssemblerFieldResponseDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fields fetched",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelFieldResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination parameters"
          },
          "404": {
            "description": "Property not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "SensorResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sensorCode": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "manufacturer": {
            "type": "string"
          },
          "installationDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "MAINTENANCE",
              "OFFLINE"
            ]
          },
          "batteryLevel": {
            "type": "integer",
            "format": "int32"
          },
          "lastMaintenance": {
            "type": "string",
            "format": "date-time"
          },
          "fieldId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
...
      "HarvestResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "harverstSeason": {
            "type": "string"
          },
          "crop": {
            "type": "string"
          },
          "plantingDate": {
            "type": "string",
            "format": "date"
          },
          "expectedHarvestDate": {
            "type": "string",
            "format": "date"
          },
          "actualStartDate": {
            "type": "string",
            "format": "date"
          },
          "actualEndDate": {
            "type": "string",
            "format": "date"
          },
          "plantedArea": {
            "type": "number"
          },
          "harvestedArea": {
            "type": "number"
          },
          "expectedYield": {
            "type": "number"
          },
          "actualYield": {
            "type": "number"
          },
          "yieldPerHectare": {
            "type": "number"
          },
          "qualityGrade": {
            "type": "string",
            "enum": [
              "PREMIUM",
              "GRADE_A",
              "GRADE_B",
              "GRADE_C",
              "REJECTED"
            ]
          },
          "marketPricePerTon": {
            "type": "number"
          },
          "totalRevenue": {
            "type": "number"
          },
          "productionCost": {
            "type": "number"
          },
          "profitMargin": {
            "type": "number"
          },
          "harvestNotes": {
            "type": "string"
          },
          "weatherConditions": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PLANNED",
              "PLANTED",
              "GROWING",
              "READY_TO_HARVEST",
              "HARVESTING",
              "COMPLETED",
              "CANCELLED"
            ]
          },
          "fieldId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
      "FieldUpdateRequestDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "crop": {
            "type": "string"
          },
          "fieldArea": {
            "type": "number",
            "minimum": 0
          },
          "areaUnit": {
            "type": "string",
            "enum": [
              "HECTARES",
              "ACRES",
              "SQUARE_METER"
            ]
          },
          "soilType": {
            "type": "string",
            "enum": [
              "CLAY",
              "SAND",
              "LOAM",
              "SILT",
              "PEAT",
              "CHALK"
            ]
          },
          "irrigationType": {
            "type": "string",
            "enum": [
              "SPRINKLER",
              "DRIP",
              "FLOOD",
              "PIVOT",
              "NONE"
            ]
          }
        }
      },
      "FieldResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "crop": {
            "type": "string"
          },
          "fieldArea": {
            "type": "number"
          },
          "areaUnit": {
            "type": "string",
            "enum": [
              "HECTARES",
              "ACRES",
              "SQUARE_METER"
            ]
          },
          "soilType": {
            "type": "string",
            "enum": [
              "CLAY",
              "SAND",
              "LOAM",
              "SILT",
              "PEAT",
              "CHALK"
            ]
          },
          "irrigationType": {
            "type": "string",
            "enum": [
              "SPRINKLER",
              "DRIP",
              "FLOOD",
              "PIVOT",
              "NONE"
            ]
          },
          "propertyId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
...
      "FieldCreateRequestDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string"
          },
          "crop": {
            "type": "string",
            "minLength": 1
          },
          "fieldArea": {
            "type": "number",
            "minimum": 0
          },
          "areaUnit": {
            "type": "string",
            "enum": [
              "HECTARES",
              "ACRES",
              "SQUARE_METER"
            ]
          },
          "soilType": {
            "type": "string",
            "enum": [
              "CLAY",
              "SAND",
              "LOAM",
              "SILT",
              "PEAT",
              "CHALK"
            ]
          },
          "irrigationType": {
            "type": "string",
            "enum": [
              "SPRINKLER",
              "DRIP",
              "FLOOD",
              "PIVOT",
              "NONE"
            ]
          },
          "propertyId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "crop",
          "fieldArea",
          "propertyId",
          "title"
        ]
      },
...
      "PageMetadata": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "PagedResourcesAssemblerSensorResponseDTO": {
        "type": "object",
        "properties": {
          "forceFirstAndLastRels": {
            "type": "boolean",
            "writeOnly": true
          }
        }
      },
      "PagedModelSensorResponseDTO": {
        "type": "object",
        "properties": {
          "_embedded": {
            "type": "object",
            "properties": {
              "sensorResponseDTOList": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/SensorResponseDTO"
                }
              }
            }
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        }
      },
...
      "PagedResourcesAssemblerHarvestResponseDTO": {
        "type": "object",
        "properties": {
          "forceFirstAndLastRels": {
            "type": "boolean",
            "writeOnly": true
          }
        }
      },
      "PagedModelHarvestResponseDTO": {
        "type": "object",
        "properties": {
          "_embedded": {
            "type": "object",
            "properties": {
              "harvestResponseDTOList": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/HarvestResponseDTO"
                }
              }
            }
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        }
      },
      "PagedResourcesAssemblerFieldResponseDTO": {
        "type": "object",
        "properties": {
          "forceFirstAndLastRels": {
            "type": "boolean",
            "writeOnly": true
          }
        }
      },
      "PagedModelFieldResponseDTO": {
        "type": "object",
        "properties": {
          "_embedded": {
            "type": "object",
            "properties": {
              "fieldResponseDTOList": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FieldResponseDTO"
                }
              }
            }
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        }
      },
      "Link": {
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "hreflang": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "deprecation": {
            "type": "string"
          },
          "profile": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "templated": {
            "type": "boolean"
          }
        }
      },
      "Links": {
        "type": "object",
        "additionalProperties": {
          "$ref": "#/components/schemas/Link"
        }
      }
    }
  }
}